refactor(opx): seperate generic functions to their own file
